Extract SpecBuilder to a separate class to keep config classes in POJO style
